+++ src/org/soulspace/overarch/adapter/template/comb.clj
@@ -12,6 +12,7 @@
             ; require to expose namespaces in templates
             [clojure.set :as set]
             [clojure.string :as str]
+            [org.soulspace.clj.string :as sstr]
             [org.soulspace.overarch.adapter.template.model-api :as m]
             [org.soulspace.overarch.adapter.template.view-api :as v]
             [org.soulspace.overarch.adapter.template.template-api :as ta]
@@ -83,12 +84,8 @@
         model-sci-ns (sci/copy-ns org.soulspace.overarch.adapter.template.model-api model-ns)
         view-ns (sci/create-ns 'org.soulspace.overarch.adapter.template.view-api)
         view-sci-ns (sci/copy-ns org.soulspace.overarch.adapter.template.view-api view-ns)
-        template-ns (sci/create-ns 'org.soulspace.overarch.adapter.template.template-api)
-        template-sci-ns (sci/copy-ns org.soulspace.overarch.adapter.template.template-api template-ns)
-        markdown-ns (sci/create-ns 'org.soulspace.overarch.adapter.template.markdown-api)
-        markdown-sci-ns (sci/copy-ns org.soulspace.overarch.adapter.template.markdown-api markdown-ns)
-        graphviz-ns (sci/create-ns 'org.soulspace.overarch.adapter.template.graphviz-api)
-        graphviz-sci-ns (sci/copy-ns org.soulspace.overarch.adapter.template.graphviz-api graphviz-ns)
+        sstr-ns (sci/create-ns 'org.soulspace.clj.string)
+        sstr-sci-ns (sci/copy-ns org.soulspace.clj.string sstr-ns)
         set-ns (sci/create-ns 'clojure.set)
         set-sci-ns (sci/copy-ns clojure.set set-ns)
         string-ns (sci/create-ns 'clojure.string)
@@ -97,16 +94,14 @@
                   'clojure.string string-sci-ns
                   'org.soulspace.overarch.adapter.template.model-api model-sci-ns
                   'org.soulspace.overarch.adapter.template.view-api view-sci-ns
-                  'org.soulspace.overarch.adapter.template.template-api template-sci-ns
-                  'org.soulspace.overarch.adapter.template.markdown-api markdown-sci-ns
-                  'org.soulspace.overarch.adapter.template.graphviz-api graphviz-sci-ns}
+                  'org.soulspace.clj.string sstr-sci-ns
+                  }
      :ns-aliases '{set clojure.set
                    str clojure.string
                    m org.soulspace.overarch.adapter.template.model-api
                    v org.soulspace.overarch.adapter.template.view-api
-                   t org.soulspace.overarch.adapter.template.template-api
-                   md org.soulspace.overarch.adapter.template.markdown-api
-                   gv org.soulspace.overarch.adapter.template.graphviz-api}
+                   sstr org.soulspace.clj.string
+                   }
      :load-fn memoized-load-fn}))
 
 (def sci-ctx (sci/init sci-opts))