+++ src/org/soulspace/overarch/application/render.clj
@@ -12,6 +12,7 @@
             [org.soulspace.overarch.domain.views.deployment-view :as deployment-view]
             [org.soulspace.overarch.domain.views.deployment-architecture-view :as deployment-architecture-view]
             [org.soulspace.overarch.domain.views.deployment-structure-view :as deployment-structure-view]
+            [org.soulspace.overarch.domain.views.domain-view :as domain-view]
             [org.soulspace.overarch.domain.views.dynamic-view :as dynamic-view]
             [org.soulspace.overarch.domain.views.glossary-view :as glossary-view]
             [org.soulspace.overarch.domain.views.model-view :as model-view]
@@ -20,8 +21,7 @@
             [org.soulspace.overarch.domain.views.state-machine-view :as state-machine-view]
             [org.soulspace.overarch.domain.views.system-landscape-view :as system-landscape-view]
             [org.soulspace.overarch.domain.views.system-structure-view :as system-structure-view]
-            [org.soulspace.overarch.domain.views.use-case-view :as use-case-view]
-            ))
+            [org.soulspace.overarch.domain.views.use-case-view :as use-case-view]))
 
 ;;;
 ;;; General rendering functions